Move error condition tests into dedicated file
