+++ Tests/ZIPFoundationTests/ZIPFoundationFileManagerTests.swift
@@ -90,7 +90,7 @@ extension ZIPFoundationTests {
             try fileManager.createParentDirectoryStructure(for: unreadableFileURL)
             let noPermissionAttributes = [FileAttributeKey.posixPermissions: Int16(0o000)]
             let result = fileManager.createFile(atPath: unreadableFileURL.path, contents: nil,
-                                                        attributes: noPermissionAttributes)
+                                                attributes: noPermissionAttributes)
             XCTAssert(result == true)
             try fileManager.zipItem(at: unreadableFileURL.deletingLastPathComponent(), to: directoryArchiveURL)
         } catch let error as CocoaError {
@@ -195,7 +195,7 @@ extension ZIPFoundationTests {
                                                             }
                                                             XCTAssert(count == pathData.count)
                                                             return pathData
-        }) else {
+                                                        }) else {
             XCTFail("Failed to read central directory structure."); return
         }
         let lfhBytes: [UInt8] = [0x50, 0x4b, 0x03, 0x04, 0x14, 0x00, 0x08, 0x08,
@@ -205,7 +205,7 @@ extension ZIPFoundationTests {
         guard let lfh = Entry.LocalFileHeader(data: Data(lfhBytes),
                                               additionalDataProvider: { _ -> Data in
                                                 return Data()
-        }) else {
+                                              }) else {
             XCTFail("Failed to read local file header."); return
         }
         guard let entry = Entry(centralDirectoryStructure: cds, localFileHeader: lfh, dataDescriptor: nil) else {
